+++ src/usb/inlretro.h
@@ -13,7 +13,7 @@ public:
 	bool open();
 
 	quint8 readByte(quint8 bank, quint16 addr, bool *ok = nullptr);
-	QByteArray readBytes(quint8 bank, quint16 addr, quint16 size, bool *ok = nullptr);
+	QByteArray readBytes(quint8 bank, quint16 addr, unsigned size, bool *ok = nullptr);
 	bool writeByte(quint8 bank, quint16 addr, quint8 data);
 
 private:
